Move schedule tests into custom folder
